% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqCounter.R
\name{SeqCounter}
\alias{SeqCounter}
\title{Counting sequence occurrences in arbitrary subsets of the genome}
\usage{
SeqCounter(genome,regions,patterns)
}
\arguments{
\item{genome}{The reference genome, BSgenome object}

\item{regions}{The genomic regions, GRanges object}

\item{patterns}{The search sequences, XStringSet object}
}
\value{
A named vector with the overall count for each pattern
}
\description{
This function, given search sequences, reference genome and genomic regions,
returns the overall count of the number of exact matches for each sequence in
the genomic regions.
}
\examples{

library(TxDb.Hsapiens.UCSC.hg38.knownGene)
library(Biostrings)

hg38 <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
promoters <- promoters(txdb, upstream=2000, downstream=200)
promoters <- trim(promoters)
patterns <- DNAStringSet(c("TATAAAA", "TATATAT", "TATAAAT"))

SeqCounter(hg38,promoters,patterns)

}
\author{
Luca Procaccio\cr Politecnico di Milano\cr Mantainer: Luca Procaccio\cr
E-mail: \href{mailto:luca.procaccio@mail.polimi.it}{luca.procaccio@mail.polimi.it}
}
